+++ test_fsic.py
@@ -31,6 +31,14 @@ import numpy as np
 import fsic
 
 
+pandas_installed = True
+
+try:
+    import pandas as pd
+except ModuleNotFoundError:
+    pandas_installed = False
+
+
 class TestRegexes(unittest.TestCase):
 
     def test_equation_re(self):
@@ -295,6 +303,34 @@ H = H[-1] + YD - C
         with self.assertRaises(fsic.DimensionError):
             # C is invalid because it has the wrong shape
             self.Model(range(10), C=[0, 0])
+
+    @unittest.skipIf(not pandas_installed, 'Requires `pandas`')
+    def test_from_dataframe(self):
+        # Test instantiation from a `pandas` DataFrame (if installed)
+        from pandas import DataFrame
+
+        # Input data with the span (index) and values
+        data = DataFrame({
+            'alpha_1': 0.6, 'alpha_2': 0.4,
+            'G': 20, 'theta': 0.2,
+        }, index=range(-5, 10))
+
+        model = self.Model.from_dataframe(data)
+
+        # Check span matches
+        self.assertEqual(model.span, list(range(-5, 10)))
+
+        # Check specified values match
+        self.assertTrue(np.allclose(model.alpha_1, 0.6))
+        self.assertTrue(np.allclose(model.alpha_2, 0.4))
+        self.assertTrue(np.allclose(model.G, 20))
+        self.assertTrue(np.allclose(model.theta, 0.2))
+
+        # All other values should be zero
+        for k in model.names:
+            if k not in ['alpha_1', 'alpha_2', 'G', 'theta']:
+                with self.subTest(variable=k):
+                    self.assertTrue(np.allclose(model[k], 0.0))
 
 
 class TestInterface(unittest.TestCase):