+++ CHANGELOG.md
@@ -11,6 +11,8 @@ Changelog](https://keepachangelog.com/en/1.0.0/), and this project adheres to
 
 ### Added
 
+- New `BaseModel` class method, `from_dataframe()`, to instantiate a model
+  object from a `pandas` DataFrame or similar.
 - Further keyword arguments to `BaseModel.solve()` are now passed to both
   `iter_periods()` and `solve_t()` e.g. to pass on arguments to over-riding
   methods in the user's own code (see